+++ packages/liveblocks-node-prosemirror/test-d/index.test-d.ts
@@ -1,0 +1,9 @@
+// import { expectType } from "tsd";
+// import { Liveblocks } from "@liveblocks/node";
+
+// const client = new Liveblocks({ secret: "sk_xxx" });
+
+(async function main() {
+  // TODO: Test withProsemirror API
+  // withProsemirror
+})();