+++ packages/liveblocks-yjs/src/index.ts
@@ -18,6 +18,7 @@ import type {
 import { ClientMsgCode, detectDupes, kInternal } from "@liveblocks/core";
 import { Base64 } from "js-base64";
 import { Observable } from "lib0/observable";
+import { IndexeddbPersistence } from "y-indexeddb";
 import * as Y from "yjs";
 
 import { Awareness } from "./awareness";
@@ -28,6 +29,7 @@ detectDupes(PKG_NAME, PKG_VERSION, PKG_FORMAT);
 
 type ProviderOptions = {
   autoloadSubdocs?: boolean;
+  offlineSupport_experimental?: boolean;
 };
 
 export class LiveblocksYjsProvider<
@@ -43,6 +45,7 @@ export class LiveblocksYjsProvider<
   private room: Room<P, S, U, E, M>;
   private rootDoc: Y.Doc;
   private options: ProviderOptions;
+  private indexeddbProvider: IndexeddbPersistence | null = null;
 
   private unsubscribers: Array<() => void> = [];
 
@@ -127,6 +130,18 @@ export class LiveblocksYjsProvider<
         this.emit("status", [this.getStatus()]);
       })
     );
+
+    if (options.offlineSupport_experimental) {
+      this.indexeddbProvider = new IndexeddbPersistence(room.id, this.rootDoc);
+      const onIndexedDbSync = () => {
+        this.rootDocHandler.synced = true;
+      };
+      this.indexeddbProvider.on("synced", onIndexedDbSync);
+
+      this.unsubscribers.push(() => {
+        this.indexeddbProvider?.off("synced", onIndexedDbSync);
+      });
+    }
 
     // different consumers listen to sync and synced
     this.rootDocHandler.on("synced", () => {
@@ -243,6 +258,11 @@ export class LiveblocksYjsProvider<
     }
     this.subdocHandlers.clear();
     super.destroy();
+  }
+
+  async clearOfflineData(): Promise<void> {
+    if (!this.indexeddbProvider) return;
+    return this.indexeddbProvider.clearData();
   }
 
   // Some provider implementations expect to be able to call connect/disconnect, implement as noop